Update Main.java to fix save/load game
Fixed an error that lied in the initialize level method, where the destroyed blocks will regenerate when the saved state is loaded.

Fixed another error, where the game would not move on to the next level if all blocks are destroyed. Happens when a saved state is loaded in. The fix was changing the destroyedBlockCount to 0.